+++ packages/moonstone/CHANGELOG.md
@@ -4,13 +4,23 @@ The following is a curated list of changes in the Enact moonstone module, newest
 
 ## [unreleased]
 
+### Added
+
+- `moonstone/VideoPlayer` props: `backwardIcon`, `forwardIcon`, `jumpBackwardIcon`, `jumpForwardIcon`, `pauseIcon`, and `playIcon` to support icon customization of the player
+- `moonstone/VideoPlayer` props `jumpButtonsDisabled` and `rateButtonsDisabled` for disabling the pairs of buttons when it's inappropriate for the playing media
+
 ### Changed
 
+- `moonstone/Slider` colors and sizing to match the latest designs
+- `moonstone/ProgressBar` to position correctly with other components nearby
 - `moonstone/Panels` breadcrumb to no longer have a horizontal line above it
 
 ### Fixed
 
+- `moonstone/Input` to not spot its own input decorator on 5way out
+- `moonstone/VideoPlayer` to no longer render its `children` in multiple places
 - `moonstone/Button` text color when used on a neutral (light) background in some cases
+- `moonstone/Popup` background opacity
 
 ## [1.0.0] - 2017-03-31